nn/requirements.txt
numpy
h5py
torch
pytorchvideo
torchvision
pyyaml
pathlib
tensorboard
tqdm
timm
packaging
einops